move integration test setup to new file
