+++ net/node.go
@@ -440,5 +440,8 @@ func newDHT(ctx context.Context, h host.Host, dsb ds.Batching) (*dualdht.DHT, er
 // Close closes the node and all its services.
 func (n Node) Close() error {
 	n.cancel()
-	return n.Peer.Close()
+	if n.Peer != nil {
+		return n.Peer.Close()
+	}
+	return nil
 }
