+++ tests/integration/utils2.go
@@ -16,12 +16,10 @@ import (
 	"fmt"
 	"os"
 	"reflect"
-	"strconv"
 	"strings"
 	"testing"
 	"time"
 
-	badger "github.com/dgraph-io/badger/v4"
 	"github.com/sourcenetwork/immutable"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -29,132 +27,11 @@ import (
 	"github.com/sourcenetwork/defradb/client"
 	"github.com/sourcenetwork/defradb/datastore"
 	badgerds "github.com/sourcenetwork/defradb/datastore/badger/v4"
-	"github.com/sourcenetwork/defradb/datastore/memory"
-	"github.com/sourcenetwork/defradb/db"
 	"github.com/sourcenetwork/defradb/errors"
 	"github.com/sourcenetwork/defradb/logging"
 	"github.com/sourcenetwork/defradb/net"
 	changeDetector "github.com/sourcenetwork/defradb/tests/change_detector"
-	"github.com/sourcenetwork/defradb/tests/clients/cli"
-	"github.com/sourcenetwork/defradb/tests/clients/http"
 )
-
-const (
-	clientGoEnvName       = "DEFRA_CLIENT_GO"
-	clientHttpEnvName     = "DEFRA_CLIENT_HTTP"
-	clientCliEnvName      = "DEFRA_CLIENT_CLI"
-	memoryBadgerEnvName   = "DEFRA_BADGER_MEMORY"
-	fileBadgerEnvName     = "DEFRA_BADGER_FILE"
-	fileBadgerPathEnvName = "DEFRA_BADGER_FILE_PATH"
-	inMemoryEnvName       = "DEFRA_IN_MEMORY"
-	mutationTypeEnvName   = "DEFRA_MUTATION_TYPE"
-)
-
-type DatabaseType string
-
-const (
-	badgerIMType   DatabaseType = "badger-in-memory"
-	defraIMType    DatabaseType = "defra-memory-datastore"
-	badgerFileType DatabaseType = "badger-file-system"
-)
-
-type ClientType string
-
-const (
-	// goClientType enables running the test suite using
-	// the go implementation of the client.DB interface.
-	goClientType ClientType = "go"
-	// httpClientType enables running the test suite using
-	// the http implementation of the client.DB interface.
-	httpClientType ClientType = "http"
-	// cliClientType enables running the test suite using
-	// the cli implementation of the client.DB interface.
-	cliClientType ClientType = "cli"
-)
-
-// The MutationType that tests will run using.
-//
-// For example if set to [CollectionSaveMutationType], all supporting
-// actions (such as [UpdateDoc]) will execute via [Collection.Save].
-//
-// Defaults to CollectionSaveMutationType.
-type MutationType string
-
-const (
-	// CollectionSaveMutationType will cause all supporting actions
-	// to run their mutations via [Collection.Save].
-	CollectionSaveMutationType MutationType = "collection-save"
-
-	// CollectionNamedMutationType will cause all supporting actions
-	// to run their mutations via their corresponding named [Collection]
-	// call.
-	//
-	// For example, CreateDoc will call [Collection.Create], and
-	// UpdateDoc will call [Collection.Update].
-	CollectionNamedMutationType MutationType = "collection-named"
-
-	// GQLRequestMutationType will cause all supporting actions to
-	// run their mutations using GQL requests, typically these will
-	// include a `id` parameter to target the specified document.
-	GQLRequestMutationType MutationType = "gql"
-)
-
-var (
-	log            = logging.MustNewLogger("tests.integration")
-	badgerInMemory bool
-	badgerFile     bool
-	inMemoryStore  bool
-	httpClient     bool
-	goClient       bool
-	cliClient      bool
-	mutationType   MutationType
-	databaseDir    string
-)
-
-const (
-	// subscriptionTimeout is the maximum time to wait for subscription results to be returned.
-	subscriptionTimeout = 1 * time.Second
-	// Instantiating lenses is expensive, and our tests do not benefit from a large number of them,
-	// so we explicitly set it to a low value.
-	lensPoolSize = 2
-)
-
-func init() {
-	// We use environment variables instead of flags `go test ./...` throws for all packages
-	//  that don't have the flag defined
-	httpClient, _ = strconv.ParseBool(os.Getenv(clientHttpEnvName))
-	goClient, _ = strconv.ParseBool(os.Getenv(clientGoEnvName))
-	cliClient, _ = strconv.ParseBool(os.Getenv(clientCliEnvName))
-	badgerFile, _ = strconv.ParseBool(os.Getenv(fileBadgerEnvName))
-	badgerInMemory, _ = strconv.ParseBool(os.Getenv(memoryBadgerEnvName))
-	inMemoryStore, _ = strconv.ParseBool(os.Getenv(inMemoryEnvName))
-
-	if value, ok := os.LookupEnv(mutationTypeEnvName); ok {
-		mutationType = MutationType(value)
-	} else {
-		// Default to testing mutations via Collection.Save - it should be simpler and
-		// faster. We assume this is desirable when not explicitly testing any particular
-		// mutation type.
-		mutationType = CollectionSaveMutationType
-	}
-
-	if !goClient && !httpClient && !cliClient {
-		// Default is to test go client type.
-		goClient = true
-	}
-
-	if changeDetector.Enabled {
-		// Change detector only uses badger file db type.
-		badgerFile = true
-		badgerInMemory = false
-		inMemoryStore = false
-	} else if !badgerInMemory && !badgerFile && !inMemoryStore {
-		// Default is to test all but filesystem db types.
-		badgerFile = false
-		badgerInMemory = true
-		inMemoryStore = true
-	}
-}
 
 // AssertPanic asserts that the code inside the specified PanicTestFunc panics.
 //
@@ -175,107 +52,6 @@ func AssertPanic(t *testing.T, f assert.PanicTestFunc) bool {
 	}
 
 	return assert.Panics(t, f, "expected a panic, but none found.")
-}
-
-func NewBadgerMemoryDB(ctx context.Context, dbopts ...db.Option) (client.DB, error) {
-	opts := badgerds.Options{
-		Options: badger.DefaultOptions("").WithInMemory(true),
-	}
-	rootstore, err := badgerds.NewDatastore("", &opts)
-	if err != nil {
-		return nil, err
-	}
-	db, err := db.NewDB(ctx, rootstore, dbopts...)
-	if err != nil {
-		return nil, err
-	}
-	return db, nil
-}
-
-func NewInMemoryDB(ctx context.Context, dbopts ...db.Option) (client.DB, error) {
-	db, err := db.NewDB(ctx, memory.NewDatastore(ctx), dbopts...)
-	if err != nil {
-		return nil, err
-	}
-	return db, nil
-}
-
-func NewBadgerFileDB(ctx context.Context, t testing.TB, dbopts ...db.Option) (client.DB, string, error) {
-	var dbPath string
-	switch {
-	case databaseDir != "":
-		// restarting database
-		dbPath = databaseDir
-
-	case changeDetector.Enabled:
-		// change detector
-		dbPath = changeDetector.DatabaseDir(t)
-
-	default:
-		// default test case
-		dbPath = t.TempDir()
-	}
-
-	opts := &badgerds.Options{
-		Options: badger.DefaultOptions(dbPath),
-	}
-	rootstore, err := badgerds.NewDatastore(dbPath, opts)
-	if err != nil {
-		return nil, "", err
-	}
-	db, err := db.NewDB(ctx, rootstore, dbopts...)
-	if err != nil {
-		return nil, "", err
-	}
-	return db, dbPath, err
-}
-
-// GetDatabase returns the database implementation for the current
-// testing state. The database type and client type on the test state
-// are used to select the datastore and client implementation to use.
-func GetDatabase(s *state) (cdb client.DB, path string, err error) {
-	dbopts := []db.Option{
-		db.WithUpdateEvents(),
-		db.WithLensPoolSize(lensPoolSize),
-	}
-
-	switch s.dbt {
-	case badgerIMType:
-		cdb, err = NewBadgerMemoryDB(s.ctx, dbopts...)
-
-	case badgerFileType:
-		cdb, path, err = NewBadgerFileDB(s.ctx, s.t, dbopts...)
-
-	case defraIMType:
-		cdb, err = NewInMemoryDB(s.ctx, dbopts...)
-
-	default:
-		err = fmt.Errorf("invalid database type: %v", s.dbt)
-	}
-
-	if err != nil {
-		return nil, "", err
-	}
-
-	switch s.clientType {
-	case httpClientType:
-		cdb, err = http.NewWrapper(cdb)
-
-	case cliClientType:
-		cdb = cli.NewWrapper(cdb)
-
-	case goClientType:
-		return
-
-	default:
-		err = fmt.Errorf("invalid client type: %v", s.dbt)
-	}
-
-	if err != nil {
-		return nil, "", err
-	}
-
-	return
 }
 
 // ExecuteTestCase executes the given TestCase against the configured database
@@ -675,12 +451,13 @@ func setStartingNodes(
 
 	// If nodes have not been explicitly configured via actions, setup a default one.
 	if !hasExplicitNode {
-		db, path, err := GetDatabase(s)
+		db, path, err := setupDatabase(s)
 		require.Nil(s.t, err)
 
-		s.nodes = append(s.nodes, &net.Node{
-			DB: db,
-		})
+		c, err := setupClient(s, &net.Node{DB: db})
+		require.Nil(s.t, err)
+
+		s.nodes = append(s.nodes, c)
 		s.dbPaths = append(s.dbPaths, path)
 	}
 }
@@ -698,16 +475,16 @@ func restartNodes(
 	for i := len(s.nodes) - 1; i >= 0; i-- {
 		originalPath := databaseDir
 		databaseDir = s.dbPaths[i]
-		db, _, err := GetDatabase(s)
+		db, _, err := setupDatabase(s)
 		require.Nil(s.t, err)
 		databaseDir = originalPath
 
 		if len(s.nodeConfigs) == 0 {
 			// If there are no explicit node configuration actions the node will be
 			// basic (i.e. no P2P stuff) and can be yielded now.
-			s.nodes[i] = &net.Node{
-				DB: db,
-			}
+			c, err := setupClient(s, &net.Node{DB: db})
+			require.NoError(s.t, err)
+			s.nodes[i] = c
 			continue
 		}
 
@@ -731,7 +508,9 @@ func restartNodes(
 			require.NoError(s.t, err)
 		}
 
-		s.nodes[i] = n
+		c, err := setupClient(s, n)
+		require.NoError(s.t, err)
+		s.nodes[i] = c
 	}
 
 	// The index of the action after the last wait action before the current restart action.
@@ -816,7 +595,7 @@ func configureNode(
 	// an in memory store.
 	cfg.Datastore.Badger.Path = s.t.TempDir()
 
-	db, path, err := GetDatabase(s) //disable change dector, or allow it?
+	db, path, err := setupDatabase(s) //disable change dector, or allow it?
 	require.NoError(s.t, err)
 
 	var n *net.Node
@@ -840,7 +619,10 @@ func configureNode(
 	s.nodeAddresses = append(s.nodeAddresses, address)
 	s.nodeConfigs = append(s.nodeConfigs, cfg)
 
-	s.nodes = append(s.nodes, n)
+	c, err := setupClient(s, n)
+	require.NoError(s.t, err)
+
+	s.nodes = append(s.nodes, c)
 	s.dbPaths = append(s.dbPaths, path)
 }
